Move results meta message inside each tab
