+++ frontend/src/components/QueryBox.less
@@ -28,36 +28,6 @@
         padding: @big-spacing;
         margin: 0;
 
-        .meta-wrapper {
-            display: table;
-            height: 100%;
-
-            .meta {
-                .big-icon {
-                    margin-right: 15px;
-                }
-
-                display: table-cell;
-                vertical-align: middle;
-
-                font-size: 14px;
-
-                &.meta-success {
-                    svg path {
-                        fill: @success;
-                    }
-                }
-
-                &.meta-error {
-                    color: @error-text;
-
-                    svg path {
-                        fill: @error-text;
-                    }
-                }
-            }
-        }
-
         .buttons-wrapper {
             text-align: right;
         }
