+++ frontend/src/components/ResultsTable.less
@@ -4,6 +4,7 @@
     flex-grow: 1;
     flex-basis: 150px;
     min-height: 150px;
+    margin-top: 2em;
 
     border: unset;
